Move shadows settings into more generic quality settings and add check for colorspace.
