+++ Assets/Digital Painting/Scripts/Validation/ValidateQaulityConfiguration.cs.meta
